refactor: extract IrisSearchAnswerService from LectureSearchService
Move WebSocket subscription, client-generated runId, and ask() logic
into a dedicated IrisSearchAnswerService. LectureSearchService now
only handles the HTTP lecture search. Update iris-answer component to
inject the new service and fix navigation-view test index assumptions.
